Changed Int64 to Int

+++ src/project_euler/problem_010.jl
@@ -23,7 +23,7 @@ problem_010(-1)         # throws DomainError
 
 Contributed by: [Praneeth Jain](https://www.github.com/PraneethJain)
 """
-function problem_010(n::Int64)
+function problem_010(n::Int)
     if n < 1
         throw(DomainError("n must be a natural number"))
     else
